fix: provide path to .d.ts typings file
Because I have moved to node 18 (or for whatever other reason),
I started to get types errors when running my tests from dependents.
Clearly, this was because types could not be found.
This change fixes it.

Note: do I need to keep my old way?
- typings field in root package.json
- package.json files in dist/[cjs|esm]/ directories
I may want to check/revisit this in the future...